API

Model Interface

ClimaCalibrate.set_up_forward_model — Function

set_up_forward_model(member, iteration, experiment_dir::AbstractString)
set_up_forward_model(member, iteration, experiment_config::ExperimentConfig)

Set up and configure a single member's forward model. Used in conjunction with run_forward_model.

This function must be overriden by a component's model interface and should set things like the parameter path and other member-specific settings.

ClimaCalibrate.run_forward_model — Function

run_forward_model(model_config)

Execute the forward model simulation with the given configuration.

This function should be overridden with model-specific implementation details. config should be obtained from set_up_forward_model: run_forward_model(set_up_forward_model(member, iter, experiment_dir))

ClimaCalibrate.observation_map — Function

observation_map(iteration)

Runs the observation map for the specified iteration. This function must be implemented for each calibration experiment.

Backend Interface

ClimaCalibrate.calibrate — Function

calibrate(::Type{JuliaBackend}, config::ExperimentConfig)
calibrate(::Type{JuliaBackend}, experiment_dir::AbstractString)

Run a calibration in Julia.

Takes an ExperimentConfig or an experiment folder. If no backend is passed, one is chosen via get_backend. This function is intended for use in a larger workflow, assuming that all needed model interface and observation map functions are set up for the calibration.

Example

Run: julia --project=experiments/surface_fluxes_perfect_model

import ClimaCalibrate

# Generate observational data and load interface
experiment_dir = dirname(Base.active_project())
include(joinpath(experiment_dir, "generate_data.jl"))
include(joinpath(experiment_dir, "observation_map.jl"))
include(joinpath(experiment_dir, "model_interface.jl"))

# Initialize and run the calibration
eki = ClimaCalibrate.calibrate(experiment_dir)

calibrate(::Type{SlurmBackend}, config::ExperimentConfig; kwargs...)
calibrate(::Type{SlurmBackend}, experiment_dir; kwargs...)

Run a full calibration, scheduling the forward model runs on Caltech's HPC cluster.

Takes either an ExperimentConfig or an experiment folder.

Keyword Arguments

`experiment_dir: Directory containing experiment configurations.

`model_interface: Path to the model interface file.

slurm_kwargs: Dictionary of slurm arguments, passed through to sbatch.

verbose::Bool: Enable verbose output for debugging.

Usage

Open julia: julia --project=experiments/surface_fluxes_perfect_model

import ClimaCalibrate: CaltechHPCBackend, calibrate

experiment_dir = dirname(Base.active_project())
model_interface = joinpath(experiment_dir, "model_interface.jl")

# Generate observational data and load interface
include(joinpath(experiment_dir, "generate_data.jl"))
include(joinpath(experiment_dir, "observation_map.jl"))
include(model_interface)

slurm_kwargs = kwargs(time = 3)
eki = calibrate(CaltechHPCBackend, experiment_dir; model_interface, slurm_kwargs);

ClimaCalibrate.sbatch_model_run — Function

sbatch_model_run(
    iter,
    member,
    output_dir,
    experiment_dir;
    model_interface,
    verbose;
    slurm_kwargs,
)

Construct and execute a command to run a forward model on a Slurm cluster for a single ensemble member.

Arguments:

iter: Iteration number

member: Member number

output_dir: Calibration experiment output directory

experiment_dir: Directory containing the experiment's Project.toml

model_interface: File containing the model interface

moduleloadstr: Commands which load the necessary modules

slurm_kwargs: Dictionary containing the slurm resources for the job. Easily generated using kwargs.

EnsembleKalmanProcesses Interface

ClimaCalibrate.initialize — Function

initialize(
    ensemble_size,
    observations,
    noise,
    prior,
    output_dir;
    rng_seed = 1234,
)
initialize(config::ExperimentConfig; rng_seed = 1234)
initialize(filepath::AbstractString; rng_seed = 1234)

Initializes the calibration process by setting up the EnsembleKalmanProcess object and parameter files with a given seed for random number generation.

ClimaCalibrate.save_G_ensemble — Function

save_G_ensemble(config::ExperimentConfig, iteration, G_ensemble)
save_G_ensemble(output_dir::AbstractString, iteration, G_ensemble)

Saves the ensemble's observation map output to the correct directory based on the provided configuration. Takes an output directory, either extracted from an ExperimentConfig or passed directly.

ClimaCalibrate.update_ensemble — Function

update_ensemble(output_dir::AbstractString, iteration, prior)
update_ensemble(config::ExperimentConfig, iteration)
update_ensemble(config_file::AbstractString, iteration)

Updates the EnsembleKalmanProcess object and saves the parameters for the next iteration.

ClimaCalibrate.ExperimentConfig — Type

ExperimentConfig(
    n_iterations::Integer,
    ensemble_size::Integer,
    observations,
    noise,
    prior::ParameterDistribution,
    output_dir,
)
ExperimentConfig(filepath::AbstractString; kwargs...)

Construct an ExperimentConfig from a given YAML file or directory containing 'experiment_config.yml'.

ExperimentConfig holds the configuration for a calibration experiment. This can be constructed from a YAML configuration file or directly using individual parameters.

ClimaCalibrate.get_prior — Function

get_prior(param_dict::AbstractDict; names = nothing)
get_prior(prior_path::AbstractString; names = nothing)

Constructs the combined prior distribution from a param_dict or a TOML configuration file specified by prior_path. If names is provided, only those parameters are used.

ClimaCalibrate.get_param_dict — Function

get_param_dict(distribution; names)

Generates a dictionary for parameters based on the specified distribution, assumed to be of floating-point type. If names is not provided, the distribution's names will be used.

ClimaCalibrate.path_to_iteration — Function

path_to_iteration(output_dir, iteration)

Creates the path to the directory for a specific iteration within the specified output directory.